Proper error message

+++ src/app/handlers/jobHandler.js
@@ -17,9 +17,13 @@ export async function handlerGetById(params) {
             where: { id: id }
         })
 
-        return NextResponse.json(job)
+        if (job) {
+            return NextResponse.json(job)
+        } else {
+            return NextResponse.json({ message: "Job not found" }, { status: 404 })
+        }
     } catch (err) {
-        return NextResponse.json(err.message, { status: 500 })
+        return NextResponse.json({ message: 'Internal server error' }, { status: 500 });
     }
 }
 
@@ -28,13 +32,13 @@ export async function handlerPost(req = NextRequest) {
         const body = await req.json()
         const { nama } = body
 
-        const newJob = await prisma.job.create({
+        await prisma.job.create({
             data: { nama }
         })
 
-        return NextResponse.json(newJob)
+        return NextResponse.json({ message: "Success create job" })
     } catch (err) {
-        return NextResponse.json(err.message, { status: 500 })
+        return NextResponse.json({ message: 'Internal server error' }, { status: 500 });
     }
 }
 
@@ -42,6 +46,16 @@ export async function handlerPut(req = NextRequest, params) {
     const id = parseInt(params.id)
 
     try {
+        const existingJob = await prisma.job.findUnique({
+            where: {
+                id: id
+            }
+        })
+
+        if (!existingJob) {
+            return NextResponse.json({ message: 'Job Not Found' }, { status: 404 });
+        }
+
         const body = await req.json()
         const { nama } = body
 
@@ -56,7 +70,7 @@ export async function handlerPut(req = NextRequest, params) {
 
         return NextResponse.json({ message: "Update job successfully" })
     } catch (err) {
-        return NextResponse.json(err.message, { status: 500 })
+        return NextResponse.json({ message: 'Internal server error' }, { status: 500 });
     }
 }
 
@@ -64,6 +78,16 @@ export async function handlerDelete(params) {
     const id = parseInt(params.id)
 
     try {
+        const existingJob = await prisma.job.findUnique({
+            where: {
+                id: id
+            }
+        })
+
+        if (!existingJob) {
+            return NextResponse.json({ message: 'Job Not Found' }, { status: 404 });
+        }
+
         await prisma.job.delete({
             where: {
                 id: id
@@ -72,6 +96,6 @@ export async function handlerDelete(params) {
 
         return NextResponse.json({ message: "Delete job successfully" })
     } catch (err) {
-        return NextResponse.json(err.message, { status: 500 })
+        return NextResponse.json({ message: 'Internal server error' }, { status: 500 });
     }
 }
